initial commit from pysisyphus dadbd5d
currently, it doesn't work with sympy 1.11, but with 1.10.

The update to sympy 1.11 forces sympficiation of the return values of
sympy.Function.eval() and all eval() methods must be classmethods.

Right now, all the Shells and 1d, as well as 3d integrals classes
inherit from sympy.Function. Some of the eval-methods are not
classmethods.

This part from the sympy docs probably applies: """If you find yourself
defining an eval() method on a Function subclass where you always return
a value and never return None, you should consider just using a normal
Python function instead, as there is no benefit to using a symbolic
Function subclass in that case (see the Best Practices for eval()
section below)"""