add: different normalization strategy
"--norm cgto" now allows a split normalization strategy, where the
normalization constant is split into a part dependent only on orbital
exponents and contraction coefficients and total angular momentum
and a second part that explicitly depends on the l, m, n in a Cartesian
shell.

The first part can be computed outside of the actual integral functions
and multiplied onto the contraction coefficients. The lmn part is
computed inside sympleints and multiplied onto the contraction
coefficients.